+++ crates/rtlola-hir/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ArgMin / ArgMax instance aggregations
 - Added Logarithmic functions
 
+### Changed
+- OutputReference now also contains a kind (parameterized or unparameterized)
+
 ## [0.6.0] - 15.05.2025
 
 ### Added